Move Testnet Settings to Experimental Features
