+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/experimental/testnet/TestnetSettingsModule.kt
@@ -1,15 +1,14 @@
-package io.horizontalsystems.bankwallet.modules.settings.experimental
+package io.horizontalsystems.bankwallet.modules.settings.experimental.testnet
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 
-object TimeLockModule {
-
+object TestnetSettingsModule {
     class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return TimeLockViewModel(App.localStorage) as T
+            return TestnetSettingsViewModel(TestnetSettingsService(App.evmTestnetManager)) as T
         }
     }
 }
