+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/experimental/testnet/TestnetSettingsService.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.settings.security.testnet
+package io.horizontalsystems.bankwallet.modules.settings.experimental.testnet
 
 import io.horizontalsystems.bankwallet.core.managers.EvmTestnetManager
