+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/experimental/timelock/TimeLockModule.kt
@@ -1,14 +1,15 @@
-package io.horizontalsystems.bankwallet.modules.settings.security.testnet
+package io.horizontalsystems.bankwallet.modules.settings.experimental.timelock
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 
-object TestnetSettingsModule {
+object TimeLockModule {
+
     class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return TestnetSettingsViewModel(TestnetSettingsService(App.evmTestnetManager)) as T
+            return TimeLockViewModel(App.localStorage) as T
         }
     }
 }
