+++ app/src/main/res/values/strings.xml
@@ -918,6 +918,8 @@
     <string name="ExperimentalFeatures_Title">Experimental</string>
     <string name="ExperimentalFeatures_Description">The features below are experimental and should be used with caution. While we have thoroughly tested these features using our own crypto funds, we cannot guarantee they will work as expected in all possible cases.</string>
     <string name="ExperimentalFeatures_BitcoinHodling">TimeLock</string>
+    <string name="ExperimentalFeatures_Testnet_Enable">Enable Testnet</string>
+    <string name="ExperimentalFeatures_Testnet_Support">Mnemonic Phrase wallets only</string>
 
     <!-- Bitcoin Hodling -->
 
@@ -1109,11 +1111,6 @@
     <string name="CoinSettings_BitcoinCashCoinType_Type145_Title">CashAddress (recommended)</string>
     <string name="CoinSettings_BitcoinCashCoinType_Type0_Description">Common before 2018</string>
     <string name="CoinSettings_BitcoinCashCoinType_Type145_Description">Common after 2018</string>
-
-    <!-- Testnet Settings -->
-
-    <string name="Testnet_Enable">Enable Testnet</string>
-    <string name="Testnet_Support">Mnemonic Phrase wallets only</string>
 
     <!-- Tor -->
